requirements.txt
Flask>=3.0,<4.0
gunicorn>=21.2,<23.0
mysql-connector-python>=9.0,<10.0
Pillow>=10.0,<12.0
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0
python-dotenv>=1.0,<2.0